requirements.txt
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kiteconnect==5.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
requests==2.32.3
service-identity==24.2.0
setuptools==80.8.0
six==1.17.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
zope.interface==7.2
